Fence auslagern
