+++ make/modules/java.base/Lib.gmk
@@ -165,49 +165,6 @@ ifeq ($(call isTargetOsType, unix), true)
 endif
 
 ################################################################################
-# Create the libjnmt library
-
-ifeq ($(call isTargetOsType, unix), true)
-  ifeq ($(STATIC_BUILD), false)
-    LIBJSIG_CFLAGS += -DHOTSPOT_VM_DISTRO='"$(HOTSPOT_VM_DISTRO)"'
-    $(eval $(call SetupJdkLibrary, BUILD_LIBJSIG, \
-        NAME := jnmt, \
-        OPTIMIZATION := LOW, \
-        CFLAGS := $(CFLAGS_JDKLIB) $(LIBJSIG_CFLAGS), \
-        LDFLAGS := $(LDFLAGS_JDKLIB) \
-            $(call SET_SHARED_LIBRARY_ORIGIN), \
-        LIBS_linux := $(LIBDL), \
-        LIBS_aix := $(LIBDL), \
-    ))
-
-    TARGETS += $(BUILD_LIBJSIG)
-
-    ############################################################################
-    # Create symlinks to libjnmt in each JVM variant sub dir
-    ifneq ($(STATIC_LIBS), true)
-      LIB_OUTPUTDIR := $(call FindLibDirForModule, java.base)
-
-      # $1 variant subdir
-      define CreateSymlinks
-        # Always symlink from libdir/variant/libjnmt.so -> ../libjnmt.so.
-        $(LIB_OUTPUTDIR)/$1/$(call SHARED_LIBRARY,jnmt): \
-            $(LIB_OUTPUTDIR)/$(call SHARED_LIBRARY,jnmt)
-		$$(call MakeDir, $$(@D))
-		$(RM) $$@
-		$(LN) -s ../$$(@F) $$@
-
-        TARGETS += $(LIB_OUTPUTDIR)/$1/$(call SHARED_LIBRARY,jnmt)
-      endef
-
-      # The subdir is the same as the variant
-      $(foreach v, $(JVM_VARIANTS), $(eval $(call CreateSymlinks,$v)))
-    endif
-    ############################################################################
-
-  endif
-endif
-
-################################################################################
 # Create the symbols file for static builds.
 
 ifeq ($(STATIC_BUILD), true)
