+++ src/hotspot/share/memory/metaspace/fence.inline.hpp
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2020, 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020 SAP SE. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ *
+ */
+
+#ifndef SHARE_MEMORY_METASPACE_FENCE_INLINE_HPP
+#define SHARE_MEMORY_METASPACE_FENCE_INLINE_HPP
+#ifdef ASSERT
+
+#include "memory/metaspace/fence.hpp"
+#include "utilities/align.hpp"
+#include "utilities/debug.hpp"
+
+namespace metaspace {
+
+STATIC_ASSERT(is_aligned(sizeof(Fence), BytesPerWord));
+
+inline void Fence::verify() const {
+  assert(_eye1 == EyeCatcher && _eye2 == EyeCatcher,
+         "Metaspace corruption: fence block at " PTR_FORMAT " broken.", p2i(this));
+}
+
+} // namespace metaspace
+
+#endif // ASSERT
+#endif // SHARE_MEMORY_METASPACE_FENCE_INLINE_HPP
+
+
+
