+++ src/hotspot/share/memory/metaspace/metaspaceArena.hpp
@@ -38,11 +38,12 @@ class Mutex;
 namespace metaspace {
 
 class ArenaGrowthPolicy;
-class ChunkManager;
-class Metachunk;
-class FreeBlocks;
-
 struct ArenaStats;
+class ChunkManager;
+class Fence;
+class FreeBlocks;
+class Metachunk;
+
 
 // The MetaspaceArena is a growable metaspace memory pool belonging to a CLD;
 //  internally it consists of a list of metaspace chunks, of which the head chunk
@@ -104,28 +105,7 @@ class MetaspaceArena : public CHeapObj<mtClass> {
   // A name for purely debugging/logging purposes.
   const char* const _name;
 
-#ifdef ASSERT
-  // Allocation guards: When active, arena allocations are interleaved with
-  //  fence allocations. An overwritten fence indicates a buffer overrun in either
-  //  the preceding or the following user block. All fences are linked together;
-  //  validating the fences just means walking that linked list.
-  // Note that for the Arena, fence blocks are just another form of user blocks.
-  class Fence {
-    static const uintx EyeCatcher =
-      NOT_LP64(0x77698465) LP64_ONLY(0x7769846577698465ULL); // "META" resp "METAMETA"
-    // Two eyecatchers to easily spot a corrupted _next pointer
-    const uintx _eye1;
-    const Fence* const _next;
-    NOT_LP64(uintx _dummy;)
-    const uintx _eye2;
-  public:
-    Fence(const Fence* next) : _eye1(EyeCatcher), _next(next), _eye2(EyeCatcher) {}
-    const Fence* next() const { return _next; }
-    void verify() const;
-  };
-  const Fence* _first_fence;
-  STATIC_ASSERT(is_aligned(sizeof(Fence), BytesPerWord));
-#endif // ASSERT
+  DEBUG_ONLY(const Fence* _first_fence;)
 
   ChunkManager* chunk_manager() const           { return _chunk_manager; }
